Refactor: Move detail screen components to a shared location
This commit relocates shared UI components from the `movie_details` and `tv_show_details` subpackages to a more generic `components` package within their respective feature modules (`movies` and `tv-shows`).

This change improves the project structure by making these components more accessible and reusable within their feature modules, removing the overly specific `_details` nesting.

Key changes:
- Moved various composables like `HeroSection`, `InfoItem`, `TrailersSection`, and `CastSection` to the new `ui/components` directory.
- Updated import statements in `MovieDetailsScreen.kt` and `TvShowDetailsScreen.kt` to reflect the new file locations.
- Corrected a few minor import paths in `MovieHeroSection.kt`, `SeriesInfoSection.kt`, and `HeroSection.kt`.
