+++ features/tv-shows/src/commonMain/kotlin/com/elna/moviedb/feature/tvshows/ui/components/CastSection.kt
@@ -1,27 +1,30 @@
-package com.elna.moviedb.feature.tvshows.ui.tv_show_details.components
+package com.elna.moviedb.feature.tvshows.ui.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.elna.moviedb.core.model.TvShowDetails
 import com.elna.moviedb.resources.Res
-import com.elna.moviedb.resources.overview
+import com.elna.moviedb.resources.cast
 import org.jetbrains.compose.resources.stringResource
 
 @Composable
-internal fun OverviewSection(tvShow: TvShowDetails) {
-    val overview = tvShow.overview
-    if (!overview.isNullOrBlank()) {
+internal fun CastSection(
+    tvShow: TvShowDetails,
+    onCastMemberClick: (Int) -> Unit
+) {
+    tvShow.cast?.takeIf { it.isNotEmpty() }?.let { cast ->
         Card(
             modifier = Modifier.fillMaxWidth(),
             elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
@@ -31,25 +34,23 @@ internal fun OverviewSection(tvShow: TvShowDetails) {
                 verticalArrangement = Arrangement.spacedBy(12.dp)
             ) {
                 Text(
-                    text = stringResource(Res.string.overview),
+                    text = stringResource(Res.string.cast),
                     style = MaterialTheme.typography.titleLarge,
                     fontWeight = FontWeight.Bold
                 )
 
-                Text(
-                    text = overview,
-                    style = MaterialTheme.typography.bodyMedium,
-                    lineHeight = MaterialTheme.typography.bodyMedium.lineHeight * 1.4
-                )
-
-                tvShow.tagline?.takeIf { it.isNotBlank() }?.let { tagline ->
-                    Text(
-                        text = "\"$tagline\"",
-                        style = MaterialTheme.typography.bodyMedium,
-                        fontWeight = FontWeight.Medium,
-                        color = MaterialTheme.colorScheme.primary,
-                        modifier = Modifier.alpha(0.8f)
-                    )
+                LazyRow(
+                    horizontalArrangement = Arrangement.spacedBy(12.dp)
+                ) {
+                    items(
+                        items = cast,
+                        key = { it.id }
+                    ) { castMember ->
+                        CastMemberCard(
+                            castMember = castMember,
+                            onClick = { onCastMemberClick(castMember.id) }
+                        )
+                    }
                 }
             }
         }
