+++ features/tv-shows/src/commonMain/kotlin/com/elna/moviedb/feature/tvshows/ui/components/GenresSection.kt
@@ -1,14 +1,14 @@
-package com.elna.moviedb.feature.tvshows.ui.tv_show_details.components
+package com.elna.moviedb.feature.tvshows.ui.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.SuggestionChip
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -16,12 +16,12 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.elna.moviedb.core.model.TvShowDetails
 import com.elna.moviedb.resources.Res
-import com.elna.moviedb.resources.trailers
+import com.elna.moviedb.resources.genres
 import org.jetbrains.compose.resources.stringResource
 
 @Composable
-internal fun TrailersSection(tvShow: TvShowDetails) {
-    tvShow.trailers?.takeIf { it.isNotEmpty() }?.let { trailers ->
+internal fun GenresSection(tvShow: TvShowDetails) {
+    tvShow.genres?.takeIf { it.isNotEmpty() }?.let { genres ->
         Card(
             modifier = Modifier.fillMaxWidth(),
             elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
@@ -31,16 +31,20 @@ internal fun TrailersSection(tvShow: TvShowDetails) {
                 verticalArrangement = Arrangement.spacedBy(12.dp)
             ) {
                 Text(
-                    text = stringResource(Res.string.trailers),
+                    text = stringResource(Res.string.genres),
                     style = MaterialTheme.typography.titleLarge,
                     fontWeight = FontWeight.Bold
                 )
 
-                LazyRow(
-                    horizontalArrangement = Arrangement.spacedBy(12.dp)
+                FlowRow(
+                    horizontalArrangement = Arrangement.spacedBy(8.dp),
+                    verticalArrangement = Arrangement.spacedBy(8.dp)
                 ) {
-                    items(trailers) { trailer ->
-                        TrailerCard(trailer = trailer)
+                    genres.forEach { genre ->
+                        SuggestionChip(
+                            onClick = { },
+                            label = { Text(text = genre) }
+                        )
                     }
                 }
             }
