+++ features/tv-shows/src/commonMain/kotlin/com/elna/moviedb/feature/tvshows/ui/components/NetworksSection.kt
@@ -1,11 +1,11 @@
-package com.elna.moviedb.feature.tvshows.ui.tv_show_details.components
+package com.elna.moviedb.feature.tvshows.ui.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
+import androidx.compose.material3.AssistChip
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.MaterialTheme
@@ -16,15 +16,12 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.elna.moviedb.core.model.TvShowDetails
 import com.elna.moviedb.resources.Res
-import com.elna.moviedb.resources.cast
+import com.elna.moviedb.resources.networks
 import org.jetbrains.compose.resources.stringResource
 
 @Composable
-internal fun CastSection(
-    tvShow: TvShowDetails,
-    onCastMemberClick: (Int) -> Unit
-) {
-    tvShow.cast?.takeIf { it.isNotEmpty() }?.let { cast ->
+internal fun NetworksSection(tvShow: TvShowDetails) {
+    tvShow.networks?.takeIf { it.isNotEmpty() }?.let { networks ->
         Card(
             modifier = Modifier.fillMaxWidth(),
             elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
@@ -34,21 +31,19 @@ internal fun CastSection(
                 verticalArrangement = Arrangement.spacedBy(12.dp)
             ) {
                 Text(
-                    text = stringResource(Res.string.cast),
+                    text = stringResource(Res.string.networks),
                     style = MaterialTheme.typography.titleLarge,
                     fontWeight = FontWeight.Bold
                 )
 
-                LazyRow(
-                    horizontalArrangement = Arrangement.spacedBy(12.dp)
+                FlowRow(
+                    horizontalArrangement = Arrangement.spacedBy(8.dp),
+                    verticalArrangement = Arrangement.spacedBy(8.dp)
                 ) {
-                    items(
-                        items = cast,
-                        key = { it.id }
-                    ) { castMember ->
-                        CastMemberCard(
-                            castMember = castMember,
-                            onClick = { onCastMemberClick(castMember.id) }
+                    networks.forEach { network ->
+                        AssistChip(
+                            onClick = { },
+                            label = { Text(text = network) }
                         )
                     }
                 }
