+++ features/tv-shows/src/commonMain/kotlin/com/elna/moviedb/feature/tvshows/ui/components/OverviewSection.kt
@@ -1,27 +1,27 @@
-package com.elna.moviedb.feature.tvshows.ui.tv_show_details.components
+package com.elna.moviedb.feature.tvshows.ui.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.SuggestionChip
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.elna.moviedb.core.model.TvShowDetails
 import com.elna.moviedb.resources.Res
-import com.elna.moviedb.resources.genres
+import com.elna.moviedb.resources.overview
 import org.jetbrains.compose.resources.stringResource
 
 @Composable
-internal fun GenresSection(tvShow: TvShowDetails) {
-    tvShow.genres?.takeIf { it.isNotEmpty() }?.let { genres ->
+internal fun OverviewSection(tvShow: TvShowDetails) {
+    val overview = tvShow.overview
+    if (!overview.isNullOrBlank()) {
         Card(
             modifier = Modifier.fillMaxWidth(),
             elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
@@ -31,21 +31,25 @@ internal fun GenresSection(tvShow: TvShowDetails) {
                 verticalArrangement = Arrangement.spacedBy(12.dp)
             ) {
                 Text(
-                    text = stringResource(Res.string.genres),
+                    text = stringResource(Res.string.overview),
                     style = MaterialTheme.typography.titleLarge,
                     fontWeight = FontWeight.Bold
                 )
 
-                FlowRow(
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                    verticalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    genres.forEach { genre ->
-                        SuggestionChip(
-                            onClick = { },
-                            label = { Text(text = genre) }
-                        )
-                    }
+                Text(
+                    text = overview,
+                    style = MaterialTheme.typography.bodyMedium,
+                    lineHeight = MaterialTheme.typography.bodyMedium.lineHeight * 1.4
+                )
+
+                tvShow.tagline?.takeIf { it.isNotBlank() }?.let { tagline ->
+                    Text(
+                        text = "\"$tagline\"",
+                        style = MaterialTheme.typography.bodyMedium,
+                        fontWeight = FontWeight.Medium,
+                        color = MaterialTheme.colorScheme.primary,
+                        modifier = Modifier.alpha(0.8f)
+                    )
                 }
             }
         }
