+++ features/tv-shows/src/commonMain/kotlin/com/elna/moviedb/feature/tvshows/ui/components/RatingsSection.kt
@@ -1,4 +1,4 @@
-package com.elna.moviedb.feature.tvshows.ui.tv_show_details.components
+package com.elna.moviedb.feature.tvshows.ui.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
