+++ features/tv-shows/src/commonMain/kotlin/com/elna/moviedb/feature/tvshows/ui/tv_show_details/TvShowDetailsScreen.kt
@@ -16,18 +16,18 @@ import androidx.compose.ui.unit.dp
 import com.elna.moviedb.core.ui.design_system.AppErrorComponent
 import com.elna.moviedb.core.ui.design_system.AppLoader
 import com.elna.moviedb.feature.tvshows.model.TvShowDetailsEvent
-import com.elna.moviedb.feature.tvshows.ui.tv_show_details.components.BasicInfoSection
-import com.elna.moviedb.feature.tvshows.ui.tv_show_details.components.CastSection
-import com.elna.moviedb.feature.tvshows.ui.tv_show_details.components.EpisodesSection
-import com.elna.moviedb.feature.tvshows.ui.tv_show_details.components.GenresSection
-import com.elna.moviedb.feature.tvshows.ui.tv_show_details.components.HeroSection
-import com.elna.moviedb.feature.tvshows.ui.tv_show_details.components.LanguagesSection
-import com.elna.moviedb.feature.tvshows.ui.tv_show_details.components.NetworksSection
-import com.elna.moviedb.feature.tvshows.ui.tv_show_details.components.OverviewSection
-import com.elna.moviedb.feature.tvshows.ui.tv_show_details.components.ProductionSection
-import com.elna.moviedb.feature.tvshows.ui.tv_show_details.components.RatingsSection
-import com.elna.moviedb.feature.tvshows.ui.tv_show_details.components.SeriesInfoSection
-import com.elna.moviedb.feature.tvshows.ui.tv_show_details.components.TrailersSection
+import com.elna.moviedb.feature.tvshows.ui.components.BasicInfoSection
+import com.elna.moviedb.feature.tvshows.ui.components.CastSection
+import com.elna.moviedb.feature.tvshows.ui.components.EpisodesSection
+import com.elna.moviedb.feature.tvshows.ui.components.GenresSection
+import com.elna.moviedb.feature.tvshows.ui.components.HeroSection
+import com.elna.moviedb.feature.tvshows.ui.components.LanguagesSection
+import com.elna.moviedb.feature.tvshows.ui.components.NetworksSection
+import com.elna.moviedb.feature.tvshows.ui.components.OverviewSection
+import com.elna.moviedb.feature.tvshows.ui.components.ProductionSection
+import com.elna.moviedb.feature.tvshows.ui.components.RatingsSection
+import com.elna.moviedb.feature.tvshows.ui.components.SeriesInfoSection
+import com.elna.moviedb.feature.tvshows.ui.components.TrailersSection
 import org.koin.compose.viewmodel.koinViewModel
 import org.koin.core.parameter.parametersOf
